// packages/lagrange-tracker/src/API.ts
export type CreateCityInput = {
  id?: string | null,
  level: number,
  submitter: string,
  type: string,
  createdAt?: string | null,
  owner?: string | null,
  cityPosId: string,
};
// ...
export type City = {
  __typename: "City",
  id: string,
  level: number,
  pos: Cordinate,
  submitter: string,
  type: string,
  createdAt: string,
  owner?: string | null,
  updatedAt: string,
  cityPosId: string,
};

export type Cordinate = {
  __typename: "Cordinate",
  id: string,
  x: number,
  y: number,
  createdAt: string,
  updatedAt: string,
};

export type UpdateCityInput = {
  id: string,
  level?: number | null,
  submitter?: string | null,
  type?: string | null,
  createdAt?: string | null,
  owner?: string | null,
  cityPosId?: string | null,
};

export type DeleteCityInput = {
  id: string,
};

export type CreateCordinateInput = {
  id?: string | null,
  x: number,
  y: number,
};
// ...
export type CreateCityMutation = {
  createCity?:  {
    __typename: "City",
    id: string,
    level: number,
    pos:  {
      __typename: "Cordinate",
      id: string,
      x: number,
      y: number,
      createdAt: string,
      updatedAt: string,
    },
    submitter: string,
    type: string,
    createdAt: string,
    owner?: string | null,
    updatedAt: string,
    cityPosId: string,
  } | null,
};
// ...
export type CreateCordinateMutation = {
  createCordinate?:  {
    __typename: "Cordinate",
    id: string,
    x: number,
    y: number,
    createdAt: string,
    updatedAt: string,
  } | null,
};